[Protocol3] Added better data encoding

+++ packages/loopring_v3/contracts/impl/libexchange/ExchangeDeposits.sol
@@ -134,7 +134,7 @@ library ExchangeDeposits
                                       // In the circuit we always overwrite the public keys in
                                       // the Account leaf with the data given onchain.
                     account.pubKeyY,
-                    tokenID,
+                    uint8(tokenID),
                     amount
                 )
             ),
